+++ src/main/aerospike/as_list.c
@@ -155,7 +155,7 @@ static bool as_list_val_tostring_foreach(as_val * val, void * udata)
 	as_list_val_tostring_data * data = (as_list_val_tostring_data *) udata;
 
 	char * str = as_val_tostring(val);
-	size_t len = strlen(str);
+	int len = (int)strlen(str);
 
 	if ( data->pos + len + 2 >= data->cap ) {
 		uint32_t adj = ((len+2) > data->blk) ? len+2 : data->blk;
